+++ config/environments/development.rb
@@ -30,12 +30,25 @@ Rails.application.configure do
   # Store uploaded files on the local file system (see config/storage.yml for options)
   config.active_storage.service = :amazon
 
-  # Don't care if the mailer can't send.
-  config.action_mailer.raise_delivery_errors = false
-  # config.action_mailer.delivery_method = :letter_opener
-  config.action_mailer.delivery_method = :sendmail
   config.action_mailer.perform_caching = false
-  config.action_mailer.default_url_options = { host: "localhost:3001" }
+  config.action_mailer.delivery_method = :smtp
+  config.action_mailer.smtp_settings = {
+    address: "email-smtp.us-east-2.amazonaws.com",
+    port: 587,
+    # domain: "customtattoodesign.ca",
+    user_name: ENV["SMTP_USER_NAME"],
+    password: ENV["SMTP_PASSWORD"],
+    authentication: "plain",
+    enable_starttls_auto: true
+  }
+  config.action_mailer.default_url_options = {
+    host: "localhost:3001"
+    # host: "api.customtattoodesign.ca",
+    # protocol: "https"
+  }
+  config.action_mailer.perform_caching = true
+
+  ####
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log