+++ lib/services/campaign_monitor.rb
@@ -24,6 +24,7 @@ module Services
           { 'Key': 'Purchased', 'Value': TaskHelper.yesno(req.deposit_order_id) },
           { 'Key': 'quote_url', 'Value': req.quote_url },
           { 'Key': 'salesperson_email', 'Value': req.salesperson&.email || Settings.emails.lee },
+          { 'Key': 'variant_price', 'Value': req.variant_price }
           # { 'Key': 'requests_statuses', 'Value': user.requests.map { |request| request.state }.join(',')} # Unused
         ]