+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [4.1.0] - 2021-07-27
+### Added
+- Variant price custom field for Campaign Monitor
+- Configure gmail labels for development/staging
+- Configure SMTP for development/staging
+### Fixed
+- Logic error for DISABLE_AUTO_QUOTE_EMAILS
+
 ## [4.0.2] - 2021-07-27
 ### Added
 - Toggle for auto quote emails via ENV["DISABLE_AUTO_QUOTE_EMAILS"]